S4 class for funGp model selection data structures

This is the formal representation of the assembly of data structures delivered by the model selection routines in the funGp package. An Xfgpm object contains the trace of an optimization process, conducted to build Gaussian process models of outstanding performance.

Main methods
fgpm_factory: structural optimization of fgpm models, creator of the "Xfgpm" class.

Plotters
plot,Xfgpm-method: plot of the evolution of the algorithm with which = "evolution" or of the absolute and relative quality of the optimized model with which = "diag".

Slots

factoryCall

Object of class "factoryCall". User call reminder.

model

Object of class "fgpm". Model selected by the heuristic structural optimization algorithm.

stat

Object of class "character". Performance measure optimized to select the model. To be set from "Q2loocv", "Q2hout".

fitness

Object of class "numeric". Value of the performance measure for the selected model.

structure

Object of class "data.frame". Structural configuration of the selected model.

log.success

Object of class "antsLog". Record of models successfully evaluated during the structural optimization. It contains the structural configuration both in data.frame and "modelCall" format, along with the fitness of each model. The models are sorted by fitness, starting with the best model in the first position.

log.crashes

Object of class "antsLog". Record of models crashed during the structural optimization. It contains the structural configuration of each model, both in data.frame and "modelCall" format.

n.solspace

Object of class "numeric". Number of possible structural configurations for the optimization instance resolved.

n.explored

Object of class "numeric". Number of structural configurations successfully evaluated by the algorithm.

details

Object of class "list". Further information about the parameters of the ant colony optimization algorithm and the evolution of the fitness along the iterations.

sIn

An object of class "matrix" containing a copy of the provided scalar inputs.

fIn

An object of class "list" containing a copy of the provided functional inputs.

sOut

An object of class "matrix" containing a copy of the provided outputs.

Useful material

Manual funGp: An R Package for Gaussian Process Regression with Scalar and Functional Inputs (doi:10.18637/jss.v109.i05 )

Author

José Betancourt, François Bachoc, Thierry Klein and Jérémy Rohmer