Analytic models for the exploration of the funGp package

Set of analytic functions that take functional variables as inputs. Since they run quickly, they can be used for testing of funGp functionalities as if they were black box computer models. They cover different situations (number of scalar inputs and complexity of the inputs-output mathematical relationship).

fgp_BB1(sIn, fIn, n.tr)

fgp_BB2(sIn, fIn, n.tr)

fgp_BB3(sIn, fIn, n.tr)

fgp_BB4(sIn, fIn, n.tr)

fgp_BB5(sIn, fIn, n.tr)

fgp_BB6(sIn, fIn, n.tr)

fgp_BB7(sIn, fIn, n.tr)

Arguments

sIn

Object with class "matrix". The scalar input points. Variables are arranged by columns and coordinates by rows.

fIn

Object with class "list". The functional inputs. Each element of the list must be a matrix containing the set of curves corresponding to one functional input.

n.tr

Object with class "numeric". The number of input points provided and correspondingly, the number of observations to produce.

Details

For all the functions, the \(d_s\) scalar inputs \(x_i\) are in the real interval \([0,\,1]\) and the \(d_f\) functional inputs \(f_i(t_i)\) are defined on the interval \([0,\,1]\). Expressions for the values are as follows.

fgp_BB1 With \(d_s = 2\) \(d_f = 2\)

   x1 * sin(x2) + x1 * mean(f1) - x2^2 * diff(range(f2))

fgp_BB2 With \(d_s = 2\) and \(d_f = 2\)

   x1 * sin(x2) + mean(exp(x1 * t1) * f1) - x2^2 * mean(f2^2 * t2)

fgp_BB3 With \(d_s = 2\) and \(d_f = 2\) is the first analytical example in Muehlenstaedt et al (2017)

   x1 + 2 * x2 + 4 * mean(t1 * f1) + mean(f2)

fgp_BB4 With \(d_s = 2\) and \(d_f = 2\) is the second analytical example in preprint of Muehlenstaedt et al (2017)

   (x2 - (5 / (4 * pi^2)) * x1^2 + (5 / pi) * x1 - 6)^2 +
       10 * (1 - (1 / (8 * pi))) * cos(x1) + 10 +
       (4 / 3) * pi * (42 * mean(f1 * (1 - t1)) +
                       pi * ((x1 + 5) / 5) + 15) * mean(t2 * f2))

fgp_BB5 With \(d_s=2\) and \(d_f=2\) is inspired by the second analytical example in final version of Muehlenstaedt et al (2017)

   (x2 - (5 / (4 * pi^2)) * x1^2 + (5 / pi) * x1 - 6)^2 +
       10 * (1 - (1 / (8 * pi))) * cos(x1) + 10 +
       (4 / 3) * pi * (42 * mean(15 * f1 * (1 - t1) - 5) +
                       pi * ((x1 + 5) / 5) + 15) * mean(15 * t2 * f2))

fgp_BB6 With \(d_s = 2\) and \(d_f = 2\) is inspired by the analytical example in Nanty et al (2016)

   2 * x1^2 + 2 * mean(f1 + t1) + 2 * mean(f2 + t2) + max(f2) + x2

fgp_BB7 With \(d_s = 5\) and \(d_f = 2\) is inspired by the second analytical example in final version of Muehlenstaedt et al (2017)

   (x2 + 4 * x3 - (5 / (4 * pi^2)) * x1^2 + (5 / pi) * x1 - 6)^2 +
       10 * (1 - (1 / (8 * pi))) * cos(x1) * x2^2 * x5^3 + 10 +
       (4 / 3) * pi * (42 * sin(x4) * mean(15 * f1 * (1 - t1) - 5) +
                       pi * (((x1 * x5 + 5) / 5) + 15) * mean(15 * t2 * f2))

Value

An object of class "matrix" with the values of the output at the specified input coordinates.

Note

The functions listed above were used to validate the functionality and stability of this package. Several tests involving all main functions, plotters and getters were run for scalar-input, functional-input and hybrid-input models. In all cases, the output of the functions were correct from the statistical and programmatic perspectives. For an example on the kind of tests performed, the interested user is referred to the introductory funGp manual (doi:10.18637/jss.v109.i05 ).

References

Muehlenstaedt, T., Fruth, J., and Roustant, O. (2017), "Computer experiments with functional inputs and scalar outputs by a norm-based approach". Statistics and Computing, 27, 1083-1097. [SC]

Nanty, S., Helbert, C., Marrel, A., Pérot, N., and Prieur, C. (2016), "Sampling, metamodeling, and sensitivity analysis of numerical simulators with functional stochastic inputs". SIAM/ASA Journal on Uncertainty Quantification, 4(1), 636-659. doi:10.1137/15M1033319