S4 class for funGp Gaussian process models

This is the formal representation of Gaussian process models within the funGp package. Gaussian process models are useful statistical tools in the modeling of complex input-output relationships.

Main methods
fgpm: creation of funGp regression models
predict: output estimation at new input points based on a funGp model
simulate: random sampling from a funGp Gaussian process model
update: modification of data and hyperparameters of a funGp model

Plotters
plotLOO: leave-one-out diagnostic plot for a funGp model
plotPreds: plot for predictions of a funGp model
plotSims: plot for simulations of a funGp model

Slots

howCalled

Object of class "modelCall". User call reminder.

type

Object of class "character". Type of model based on type of inputs. To be set from "scalar", "functional", "hybrid".

ds

Object of class "numeric". Number of scalar inputs.

df

Object of class "numeric". Number of functional inputs.

f_dims

Object of class "numeric". An array with the original dimension of each functional input.

sIn

Object of class "matrix". The scalar input points. Variables are arranged by columns and coordinates by rows.

fIn

Object of class "list". The functional input points. Each element of the list contains a functional input in the form of a matrix. In each matrix, curves representing functional coordinates are arranged by rows.

sOut

Object of class "matrix". The scalar output values at the coordinates specified by sIn and/or fIn.

n.tot

Object of class "integer". Number of observed points used to compute the training-training and training-prediction covariance matrices.

n.tr

Object of class "integer". Among all the points loaded in the model, the amount used for training.

f_proj

Object of class "fgpProj". Data structures related to the projection of functional inputs. Check fgpProj for more details.

kern

Object of class "fgpKern". Data structures related to the kernel of the Gaussian process model. Check fgpKern for more details.

nugget

Object of class "numeric". Variance parameter standing for the homogeneous nugget effect.

preMats

Object of class "list". L and LInvY matrices pre-computed for prediction. L is a lower diagonal matrix such that \(L'L\) equals the training auto-covariance matrix \(K.tt\). On the other hand, \(LInvY = L^(-1) * sOut\).

Useful material

Manual Gaussian Process Regression for Scalar and Functional Inputs with funGp - The in-depth tour