Gaussian Process Models for Scalar and Functional Inputs

Construction and smart selection of Gaussian process models for analysis of computer experiments with emphasis on treatment of functional inputs that are regularly sampled. Smart selection is based on Ant Colony Optimization ACO algorithm.

Note

This package was first developed in the frame of the RISCOPE research project, funded by the French Agence Nationale de la Recherche (ANR) for the period 2017-2021 (ANR, project No. 16CE04-0011, RISCOPE.fr), and certified by SAFE Cluster.

Base functionalities

Main methods
fgpm: creation of funGp regression models
predict,fgpm-method: output estimation at new input points based on a funGp model
simulate,fgpm-method: random sampling from a funGp Gaussian process model
update,fgpm-method: modification of data and hyperparameters of a funGp model

Plotters
plot,fgpm-method: validation plot for a fgpm model
plot.predict.fgpm: plot of predictions based on a fgpm model
plot.simulate.fgpm: plot of simulations based on a fgpm model

Model selection

Main method
fgpm_factory: structural parameter optimization

Functions for pre-optimization
decay: regularized initial pheromones
decay2probs: normalized initial pheromones

Plotters post-optimization
plot,Xfgpm-method: plot of the evolution of the algorithm with which = "evolution" or of the absolute and relative quality of the optimized model with which = "diag"

Correction post-optimization of input data structures
which_on: indices of active inputs in a model structure delivered by fgpm_factory
get_active_in: extraction of active input data based on a model structure delivered by fgpm_factory

Useful material

Manual Gaussian Process Regression for Scalar and Functional Inputs with funGp - The in-depth tour

Paper Gaussian process metamodeling of functional-input code for coastal flood hazard assessment

Tech. report Ant Colony Based Model Selection for Functional-Input Gaussian Process Regression

Authors

José Betancourt, François Bachoc and Thierry Klein

Contributors

Déborah Idier and Jérémy Rohmer