S4 class for structures linked to projections in a funGp model

This is the formal representation for data structures linked to projection of inputs in a Gaussian process model within the funGp package.

Slots

pdims

Object of class "numeric". Projection dimension of each input.

basType

Object of class "character". To be chosen from "PCA", "B-splines".

basis

Object of class "list". Projection basis. For functional inputs, each element (fDims_i x fpDims_i) contains the basis functions used for the projection of one functional input.

coefs

Object of class "list". Each element (n x fpDims_i) contains the coefficients used for the projection of one functional input.

Author

José Betancourt, François Bachoc and Thierry Klein