Summary method for Xfgpm objects

Display a summary of the structure of a Xfgpm object, with a short description of up to n fgpm objects visited during the ACO optimization.

# S4 method for Xfgpm
summary(object, n = 24, ...)

Arguments

object

A Xfgpm object.

n

Maximal number of lines (fgpm objects) to show.

...

Not used yet.

Value

An object inheriting from list, actually a list containing one or two data frames depending on the number of inputs. In each data frame, the n rows provide information on the best fgpm objects visited.

Details

The displayed information depends on the number of candidate inputs, in order to maintain compact tables. The inputs are labelled with integer suffixes, the prefix being "X" for scalar inputs and "F" for functional inputs.

With a small number of inputs, the list contains only one data frame. For each candidate input (either scalar or functional) a column with the input name indicates if the input is active (cross x) or not (white space) in the fgpm object corresponding to the row. For each functional variable also shown are: the distance used D_, the dimension Bas_ after dimension reduction, the type of basis used B_. Remind that the kernel (Kern) is the same for all functional inputs. Also shown is the value of the Leave-One-Out coefficient Q².

With a large number of inputs, the list contains two data frames. The first one tells which inputs are active among the scalar and functional candidate inputs. The second data frame gives more details for functional inputs as before.

Examples

summary(xm)
#> Inputs and details 
#>    X1 X2 X3 X4 X5 F1 D_F1 Dim_F1 Bas_F1 F2 D_F2 Dim_F2 Bas_F2  Kern    Q2
#> 1      x  x  x  x  x  idx      1   Bspl  x  idx      3   Bspl gauss 0.939
#> 2      x  x  x  x  x  idx      1    PCA  x  idx      3   Bspl gauss 0.854
#> 3      x     x  x  x  idx      1   Bspl  x  idx      3   Bspl gauss 0.854
#> 4      x     x  x  x  grp      1   Bspl  x  idx      3   Bspl gauss 0.854
#> 5      x  x  x  x  x  idx      1   Bspl  x  idx      3   Bspl mat32 0.845
#> 6      x  x  x  x      --      -     --  x  idx      3   Bspl gauss 0.841
#> 7      x     x  x  x  idx      1   Bspl  x  idx      3   Bspl mat32 0.833
#> 8      x     x  x  x  idx      1   Bspl  x  idx      3   Bspl mat52 0.809
#> 9      x     x  x  x  idx      1   Bspl  x  grp     22    PCA gauss 0.788
#> 10     x     x  x  x  idx      1    PCA  x  idx      3   Bspl mat32 0.785
#> 11     x     x  x      --      -     --  x  idx      3   Bspl mat32 0.753
#> 12     x     x  x  x  idx      1   Bspl  x  idx      3    PCA gauss 0.721
#> 13     x  x  x  x  x  idx      1   Bspl  x  idx      1   Bspl gauss 0.678
#> 14     x  x  x         --      -     --  x  idx      3   Bspl gauss 0.623
#> 15     x     x  x  x  idx      1    PCA  x  idx      3    PCA gauss 0.596
#> 16     x     x         --      -     --  x  idx      3    PCA gauss 0.589
#> 17           x  x  x  idx      1   Bspl  x  idx      3   Bspl mat32 0.570
#> 18     x  x  x  x  x  idx      1   Bspl  x  idx      3    PCA gauss 0.553
#> 19     x     x  x  x  grp      5   Bspl  x  idx      3    PCA mat32 0.487
#> 20     x  x            --      -     --  x  idx      3   Bspl mat32 0.477
#> 21     x     x         --      -     --  x  idx      3   Bspl mat52 0.469
#> 22     x  x     x  x  grp      2    PCA  x  idx      3   Bspl mat32 0.402
#> 23     x     x         --      -     --  x  idx      3    PCA mat32 0.379
#> 24  x     x  x  x      --      -     --  x  idx      3   Bspl mat32 0.098