Summary method for fgpm objects

Display the structure of a fgpm object and the value of the parameters (variance and length-scales).

# S4 method for fgpm
summary(object, ...)

Arguments

object

An fgpm object.

...

Not used yet.

Note

This method is actually identical to the show method for this class which is called when the name of the object is entered in an interactive session.

Examples

m <- xm@model
class(m)
#> [1] "fgpm"
#> attr(,"package")
#> [1] "funGp"
summary(m)
#> 
#> Gaussian Process Model____________________________________
#> * Scalar inputs: 4
#> * Functional inputs: 2
#> 
#> | Input | Orig. dim | Proj. dim |   Basis   |  Distance  |
#> |:-----:|:---------:|:---------:|:---------:|:----------:|
#> |  F1   |    10     |     1     | B-splines | L2_byindex |
#> |  F2   |    22     |     3     | B-splines | L2_byindex |
#> 
#> * Total data points: 32
#> * Trained with: 32
#> * Kernel type: gauss
#> * Convergence: 0
#> * NegLogLik: 194.8887
#> * Hyperparameters:
#>   -> variance: 193630.7962
#>   -> length-scale:
#> 	 ls(X1): 2.0000
#> 	 ls(X2): 2.0000
#> 	 ls(X3): 1.8137
#> 	 ls(X4): 2.0000
#> 	 ls(F1): 0.6538
#> 	 ls(F2): 1.3678
#> 	 ls(F2): 1.9673
#> 	 ls(F2): 1.2500
#> __________________________________________________________ 
m
#> 
#> Gaussian Process Model____________________________________
#> * Scalar inputs: 4
#> * Functional inputs: 2
#> 
#> | Input | Orig. dim | Proj. dim |   Basis   |  Distance  |
#> |:-----:|:---------:|:---------:|:---------:|:----------:|
#> |  F1   |    10     |     1     | B-splines | L2_byindex |
#> |  F2   |    22     |     3     | B-splines | L2_byindex |
#> 
#> * Total data points: 32
#> * Trained with: 32
#> * Kernel type: gauss
#> * Convergence: 0
#> * NegLogLik: 194.8887
#> * Hyperparameters:
#>   -> variance: 193630.7962
#>   -> length-scale:
#> 	 ls(X1): 2.0000
#> 	 ls(X2): 2.0000
#> 	 ls(X3): 1.8137
#> 	 ls(X4): 2.0000
#> 	 ls(F1): 0.6538
#> 	 ls(F2): 1.3678
#> 	 ls(F2): 1.9673
#> 	 ls(F2): 1.2500
#> __________________________________________________________