S4 class for funGp model selection data structures

This is the formal representation of the assembly of data structures delivered by the model selection routines in the funGp package. Gaussian process models are useful statistical tools in the modeling of complex input-output relationships. An Xfgpm object contains the trace of an optimization process, conducted to build Gaussian process models of outstanding performance.

Main methods
fgpm_factory: structural optimization of funGp models

Plotters
plotX: diagnostic plots for a fgpm_factory optimization and the selected model
plotEvol: plot of the evolution of the model selection algorithm in funGp

Slots

factoryCall

Object of class "factoryCall". User call reminder.

model

Object of class "fgpm". Model selected by the heuristic structural optimization algorithm.

stat

Object of class "character". Performance measure optimized to select the model. To be set from "Q2loocv", "Q2hout".

fitness

Object of class "numeric". Value of the performance measure for the selected model.

structure

Object of class "data.frame". Structural configuration of the selected model.

log.success

Object of class "antsLog". Record of models successfully evaluated during the structural optimization. It contains the structural configuration both in data.frame and "modelCall" format, along with the fitness of each model. The models are sorted by fitness, starting with the best model in the first position.

log.crashes

Object of class "antsLog". Record of models crashed during the structural optimization. It contains the structural configuration of each model, both in data.frame and "modelCall" format.

n.solspace

Object of class "numeric". Number of possible structural configurations for the optimization instance resolved.

n.explored

Object of class "numeric". Number of structural configurations successfully evaluated by the algorithm.

details

Object of class "list". Further information about the parameters of the ant colony optimization algorithm and the evolution of the fitness along the iterations.

Useful material

Manual Gaussian Process Regression for Scalar and Functional Inputs with funGp - The in-depth tour