Gaussian Process Models for Scalar and Functional Inputs

Construction and smart selection of Gaussian process models with emphasis on treatment of functional inputs.

Note

This package was first developed in the frame of the RISCOPE research project, funded by the French Agence Nationale de la Recherche (ANR) for the period 2017-2021 (ANR, project No. 16CE04-0011, RISCOPE.fr), and certified by SAFE Cluster.

Base functionalities

Main methods
fgpm: creation of funGp regression models
predict: output estimation at new input points based on a funGp model
simulate: random sampling from a funGp Gaussian process model
update: modification of data and hyperparameters of a funGp model

Plotters
plotLOO: validation plot for a funGp model
plotPreds: plot of predictions based on a funGp model
plotSims: plot of simulations based on a funGp model

Model selection

Main methods
fgpm_factory: fill

Plotters pre-optimization
decay: only title
decay2probs: only title

Plotters post-optimization
plotX: only title
plotEvol: only title

Useful material

Manual Gaussian Process Regression for Scalar and Functional Inputs with funGp - The in-depth tour

Paper Gaussian process metamodeling of functional-input code for coastal flood hazard assessment

Tech. report Ant Colony Based Model Selection for Functional-Input Gaussian Process Regression

Authors

José Betancourt, François Bachoc and Thierry Klein

Contributors

Déborah Idier and Jérémy Rohmer